`indicators.rs` 和 `series.rs` 中調整測試代碼以使用新的 `DailyOhlcv` 和 `MinuteOhlcv` 類型，改進過濾和數據訪問邏輯。更新測試以反映新的 `FinancialSeries` 結構，並修正相關的類型轉換問題。

+++ src/domain_types/series.rs
@@ -93,8 +93,8 @@ impl<F: FrequencyMarker, D: DataFormat> FinancialSeries<F, D> {
     pub fn filter_date_range(self, start_time: i64, end_time: i64) -> Self {
         let filtered = self.lazy_frame.filter(
             col(ColumnName::TIME)
-                .gt_eq(lit(start_time))
-                .and(col(ColumnName::TIME).lt_eq(lit(end_time))),
+                .gt_eq(lit(start_time).cast(DataType::Int64))
+                .and(col(ColumnName::TIME).lt_eq(lit(end_time).cast(DataType::Int64))),
         );
 
         Self {
